# PERF: creating string Series/Arrays from sequence with many strings