@topper-123 2020-09-12
There are ASVs for both `str`and `"string"`in `asv_bench/benchmarks/strings.py`.